pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[project]
name = "playlab-whatsapp-bridge"
version = "0.1.0"
description = "WhatsApp bridge server for Playlab.ai"
requires-python = ">=3.10"
dependencies = [
  "fastapi",
  "uvicorn",
  "httpx",
  "sqlalchemy",
  "asyncpg",
  "greenlet",
  "alembic",
  "redis",
  "python-dotenv",
  "python-multipart",
  "anthropic",
]

[project.optional-dependencies]
test = [
  "pytest",
  "pytest-asyncio",
  "aiosqlite",
]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I"]
